migrations

+++ database/migrations/2021_01_26_210502_update_file_denylist_to_json.php
@@ -1,5 +1,6 @@
 <?php
 
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -11,13 +12,17 @@ class UpdateFileDenylistToJson extends Migration
      */
     public function up(): void
     {
-        Schema::table('eggs', function (Blueprint $table) {
-            $table->dropColumn('file_denylist');
-        });
+        if (Schema::hasColumn('eggs', 'file_denylist')) {
+            Schema::table('eggs', function (Blueprint $table) {
+                $table->dropColumn('file_denylist');
+            });
+        }
 
-        Schema::table('eggs', function (Blueprint $table) {
-            $table->json('file_denylist')->nullable()->after('docker_images');
-        });
+        if (!Schema::hasColumn('eggs', 'file_denylist')) {
+            Schema::table('eggs', function (Blueprint $table) {
+                $table->json('file_denylist')->nullable()->after('docker_images');
+            });
+        }
     }
 
     /**
@@ -25,12 +30,26 @@ class UpdateFileDenylistToJson extends Migration
      */
     public function down(): void
     {
-        Schema::table('eggs', function (Blueprint $table) {
-            $table->dropColumn('file_denylist');
-        });
+        if (Schema::hasColumn('eggs', 'file_denylist')) {
+            Schema::table('eggs', function (Blueprint $table) {
+                $table->dropColumn('file_denylist');
+            });
+        }
 
-        Schema::table('eggs', function (Blueprint $table) {
-            $table->text('file_denylist')->after('docker_images');
-        });
+        if (DB::getDriverName() === 'sqlite') {
+            if (!Schema::hasColumn('eggs', 'file_denylist')) {
+                Schema::table('eggs', function (Blueprint $table) {
+                    $table->text('file_denylist')->nullable()->after('docker_images');
+                });
+            }
+
+            return;
+        }
+
+        if (!Schema::hasColumn('eggs', 'file_denylist')) {
+            Schema::table('eggs', function (Blueprint $table) {
+                $table->text('file_denylist')->after('docker_images');
+            });
+        }
     }
 }
